// main/script/client/GoodsExchange.d.ts
//
// Autogenerated by Thrift Compiler (0.11.0)
//
// DO NOT EDIT UNLESS YOU ARE SURE THAT YOU KNOW WHAT YOU ARE DOING
//


export default class GoodsExchangeClient {
  input: Thrift.TJSONProtocol;
  output: Thrift.TJSONProtocol;
  seqid: number;

  constructor(input: Thrift.TJSONProtocol, output?: Thrift.TJSONProtocol);

  exchange(uid: number, gameId: number, goodsId: number, addressId: number): boolean;

  exchange(uid: number, gameId: number, goodsId: number, addressId: number, callback: (data: boolean)=>void): void;

  getExchangeGoods(gameId: number): string;

  getExchangeGoods(gameId: number, callback: (data: string)=>void): void;
}